Input: fix keyname logging
GetKeyName got somewhat broken in the last 15 years, fix this
so that keys are properly logged:

Log the full 16-bit code (as it would appear in key id="..." mappings)
for unknown keys instead of only the lower 8 bits and display the
obc code only for valid keycodes (255 or less).

Also handle the special KEY_UNICODE and KEY_INVALID values.

+++ xbmc/input/InputManager.cpp
@@ -574,7 +574,7 @@ bool CInputManager::HandleKey(const CKey& key)
   if (appPower->WakeUpScreenSaverAndDPMS(processKey) && !processKey)
   {
     CLog::LogF(LOGDEBUG, "{} pressed, screen saver/dpms woken up",
-               m_Keyboard.GetKeyName((int)key.GetButtonCode()));
+               m_Keyboard.GetKeyName(key.GetButtonCode()));
     return true;
   }
 
@@ -657,7 +657,7 @@ bool CInputManager::HandleKey(const CKey& key)
       }
 
       CLog::LogF(LOGDEBUG, "{} pressed, trying keyboard action {:x}",
-                 m_Keyboard.GetKeyName((int)key.GetButtonCode()), action.GetID());
+                 m_Keyboard.GetKeyName(key.GetButtonCode()), action.GetID());
 
       if (g_application.OnAction(action))
         return true;
@@ -673,7 +673,7 @@ bool CInputManager::HandleKey(const CKey& key)
   }
   if (!key.IsAnalogButton())
     CLog::LogF(LOGDEBUG, "{} pressed, window {}, action is {}",
-               m_Keyboard.GetKeyName((int)key.GetButtonCode()), iWin, action.GetName());
+               m_Keyboard.GetKeyName(key.GetButtonCode()), iWin, action.GetName());
 
   return ExecuteInputAction(action);
 }
